+++ fixtures/public-assets/src/App.tsx
