+++ fixtures/public-assets/src/client.tsx
@@ -1,6 +1,6 @@
 import { hydrateRoot } from 'react-dom/client';
 
-import App from './app.tsx';
+import App from './App.tsx';
 
 export default () => {
   hydrateRoot(document.getElementById('app')!, <App />);
