+++ fixtures/public-assets/src/render.tsx
@@ -2,7 +2,7 @@ import html from 'dedent';
 import { renderToString } from 'react-dom/server';
 import type { Render } from 'sku';
 
-import App from './app.tsx';
+import App from './App.tsx';
 
 export default {
   renderApp: ({ SkuProvider }) =>
